requirements.txt
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
ConfigArgParse==1.5.3
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.8.0
geventhttpclient==1.5.2
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
locust==2.2.3
MarkupSafe==2.0.1
msgpack==1.0.2
psutil==5.8.0
psycogreen==1.0.2
psycopg2-binary==2.9.1
python-dotenv==0.19.1
pyzmq==22.3.0
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.26.0
roundrobin==0.0.2
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
